+++ src/main/java/com/tensions/cafembtitest/repository/CafeMbtiMapper.java
@@ -7,5 +7,5 @@ import org.apache.ibatis.annotations.Param;
 public interface CafeMbtiMapper {
     void setMbtiResult(@Param("result") String result);
 
-    int getMbtiPercent(@Param("result") String result);
+    float getMbtiPercent(@Param("result") String result);
 }
