+++ src/main/java/com/tensions/cafembtitest/service/CafeMbtiService.java
@@ -5,5 +5,5 @@ import com.tensions.cafembtitest.dto.CafeMbtiDTO;
 public interface CafeMbtiService {
     String mbtiResult(CafeMbtiDTO cafeMbtiDTO);
 
-    int mbtiPercent(String result);
+    float mbtiPercent(String result);
 }
